Refact: 스팀 게임 목록 조회하는 코드를 user.controller로 이동 Refact: 회원가입 부분에서 오류 발생 시 에러를 반환하도록 처리
