+++ src/user/user.controller.ts
@@ -26,7 +26,9 @@ import { UserTitleService } from "./service/user-title.service";
 import { TokenPayload } from "src/auth/object/token-payload.obj";
 import { UploadProfileImageDto } from "./dto/upload-profile-image.dto";
 import { UpdateIntroductionDto } from "./dto/update-introduction.dto";
-import { HttpStatusCode } from "axios";
+import axios, { HttpStatusCode } from "axios";
+import { rhythmGameList } from "../auth/object/auth.object";
+import { CodecService } from "src/codec/codec.service";
 
 @Controller("user")
 export class UserController {
@@ -34,6 +36,7 @@ export class UserController {
     private readonly userService: UserService,
     private readonly plateSettingService: PlateSettingService,
     private readonly userTitleService: UserTitleService,
+    private readonly codecService: CodecService,
   ) {}
 
   @ApiTags("User Setting")
@@ -100,6 +103,7 @@ export class UserController {
     res.status(HttpStatusCode.Ok).send(titles);
   }
 
+  @ApiTags("User Setting")
   @Post("upload/profile-image")
   async uploadProfileImage(
     @Req() req,
@@ -109,10 +113,37 @@ export class UserController {
     const user: TokenPayload = req.user;
     const imageBase64: string = profileImageDto.image;
     this.userService.uploadUserProfileImage(user.nickname, imageBase64);
-    this.userService.fetchPlateData(user.uid);
     res.status(HttpStatusCode.Ok).send();
   }
 
+  @SkipAuth()
+  @ApiTags("User Setting")
+  @Get("plate/:nickname")
+  async fetchUserPlateData(
+    @Req() req,
+    @Res() res: Response,
+    @Param("nickname") nickname: string,
+  ) {
+    await this.userService.fetchPlateDataByNickname(nickname);
+    res.send();
+  }
+
+  @ApiBody({
+    schema: {
+      type: "object",
+      properties: {
+        introduction: {
+          type: "string",
+          example: "rgback admin입니다.",
+          description: "업데이트 될 한줄 소개",
+        },
+      },
+    },
+  })
+  @ApiOkResponse({
+    description: "가져오기 성공",
+  })
+  @ApiTags("User Setting")
   @Get("introduction")
   async fetchIntroduction(@Req() req, @Res() res: Response) {
     const user: TokenPayload = req.user;
@@ -120,7 +151,6 @@ export class UserController {
     res.status(HttpStatusCode.Ok).send({ introduction: introduction });
   }
 
-  @Post("introduction")
   @ApiBody({
     schema: {
       type: "object",
@@ -143,6 +173,7 @@ export class UserController {
     description: "데이터베이스 오류",
   })
   @ApiTags("User Setting")
+  @Post("introduction")
   async changeIntroduction(
     @Body() updateDto: UpdateIntroductionDto,
     @Req() req,
@@ -163,5 +194,59 @@ export class UserController {
     } else {
       res.status(HttpStatusCode.InternalServerError).send();
     }
+  }
+
+  @ApiTags("steam-games")
+  @ApiOkResponse({
+    description: "스팀 게임 목록 조회 완료",
+    schema: {
+      type: "array",
+      items: {
+        type: "string",
+      },
+      example: [
+        "Djmax Respect V",
+        "EZ2ON REBOOT",
+        "불과 얼음의 춤 (A Dance of Fire and Ice)",
+        "Rhythm Doctor",
+      ],
+    },
+  })
+  @ApiBadRequestResponse({
+    description: "저장된 스팀 id 없음",
+  })
+  @Get("steam/games")
+  async getGames(@Req() req, @Res() res: Response) {
+    const token: TokenPayload = req.user;
+    const user = await this.userService.fetchWithUserId(+token.uid);
+    const steamid: string = user.steamId;
+    if (steamid === "" || steamid == null) {
+      res.status(HttpStatusCode.BadRequest).send();
+      return;
+    }
+
+    const decrypted = await this.codecService.decrypt(steamid);
+
+    const params = {
+      key: process.env.STEAM_API_KEY,
+      steamid: decrypted,
+      format: "json",
+      include_appinfo: true,
+      appids_filter: rhythmGameList,
+    };
+
+    const games = await axios.get(
+      `https://api.steampowered.com/IPlayerService/GetOwnedGames/v0001/`,
+      { params },
+    );
+    const gamesObj: Array<Record<string, string>> = games.data.response.games;
+
+    const data = [];
+    gamesObj.forEach((game) => {
+      data.push(game.name);
+    });
+
+    res.status(HttpStatusCode.Ok).send(data);
+    return;
   }
 }
